+++ src/deluge/dsp/filter/filter.h
@@ -81,7 +81,7 @@ public:
 
 		//this is 1q31*1q16/(1q16+tan(f)/2)
 		//tan(f) is q17
-		divideBy1PlusTannedFrequency = (q31_t)(288230376151711744.0f / (float)(ONE_Q16 + (tannedFrequency >> 1)));
+		divideBy1PlusTannedFrequency = (q31_t)(288230376151711744.0 / (double)(ONE_Q16 + (tannedFrequency >> 1)));
 		fc = multiply_32x32_rshift32_rounded(tannedFrequency, divideBy1PlusTannedFrequency) << 4;
 	}
 	q31_t fc;
