+++ src/deluge/model/clip/audio_clip.cpp
@@ -534,7 +534,7 @@ void AudioClip::render(ModelStackWithTimelineCounter* modelStack, int32_t* outpu
 	//if ((sampleLengthInSamples >> 5) > clipLengthInSamples) return false;
 
 	uint64_t requiredSpeedAdjustment =
-	    (uint64_t)((float)((uint64_t)sampleLengthInSamples << 24)) / (float)clipLengthInSamples;
+	    (uint64_t)((double)((uint64_t)sampleLengthInSamples << 24)) / (double)clipLengthInSamples;
 
 	// If we're squishing time...
 	if (sampleControls.pitchAndSpeedAreIndependent) {
